pyproject.toml
[project]
name = "docfx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "pyyaml>=6.0.3",
    "ruff>=0.14.10",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.coverage.run]
omit = [
    "dev.py",
    "main.py",
]

[tool.mypy]
mypy_path = "."
explicit_package_bases = true
exclude = ["^tests/"]
